+++ src/liblogic_tile_has_box.cpp
@@ -3,7 +3,7 @@
 entityid tile_has_box(shared_ptr<gamestate> g, int x, int y, int z) {
     massert(g, "gamestate is NULL");
     massert(z >= 0, "floor is out of bounds");
-    massert(z < g->dungeon->floors->size(), "floor is out of bounds");
+    massert((size_t)z < g->dungeon->floors->size(), "floor is out of bounds");
     shared_ptr<dungeon_floor_t> df = d_get_floor(g->dungeon, z);
     shared_ptr<tile_t> t = df_tile_at(df, (vec3){x, y, z});
     for (int i = 0; (size_t)i < t->entities->size(); i++) {
