+++ src/liblogic_tile_npc_living_count.cpp
@@ -4,7 +4,7 @@
 int tile_npc_living_count(shared_ptr<gamestate> g, int x, int y, int z) {
     massert(g, "gamestate is NULL");
     massert(z >= 0, "floor is out of bounds");
-    massert(z < g->dungeon->floors->size(), "floor is out of bounds");
+    massert((size_t)z < g->dungeon->floors->size(), "floor is out of bounds");
     shared_ptr<dungeon_floor_t> df = d_get_floor(g->dungeon, z);
     massert(df, "failed to get dungeon floor");
     shared_ptr<tile_t> t = df_tile_at(df, (vec3){x, y, z});
